Fix Hal to handle port = 0 as such and not interpret it as "any" (#16280)

+++ nic/sdk/include/sdk/l4.hpp
@@ -6,8 +6,8 @@
 #include "include/sdk/types.hpp"
 
 typedef struct port_range_s {
-    uint16_t    port_lo;
-    uint16_t    port_hi;
+    uint32_t    port_lo;
+    uint32_t    port_hi;
 } __PACK__ port_range_t;
 
 #endif    // __L4_HPP__
